deployment test

+++ websocketServerModule/src/sprout/clipcon/server/controller/Server.java
@@ -7,8 +7,10 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Map;
 import java.util.Random;
+import java.util.Scanner;
 import java.util.Set;
 
 import sprout.clipcon.server.model.Group;
@@ -18,14 +20,16 @@ public class Server {
 	/** All groups on the server */
 	private Map<String, Group> groups = Collections.synchronizedMap(new HashMap<String, Group>());
 
-	 // public static final String SERVER_ROOT_LOCATION = "C:\\Users\\Administrator\\Desktop\\";
+	// public static final String SERVER_ROOT_LOCATION =
+	// "C:\\Users\\Administrator\\Desktop\\";
 	public static final String SERVER_ROOT_LOCATION = "C:\\Users\\delf\\Desktop\\";
-	
+
 	// 업로드 파일을 저장할 위치
-	 public static final String RECEIVE_LOCATION = SERVER_ROOT_LOCATION + "clipcon_download\\"; 
+	public static final String RECEIVE_LOCATION = SERVER_ROOT_LOCATION + "clipcon_download\\";
 
 	// change source
 	private Server() {
+		new ServerStatusCheck().start();
 	}
 
 	public static Server getInstance() {
@@ -43,14 +47,14 @@ public class Server {
 	 * @return 그룹의 존재 여부. 그룹이 존재하면 true, 그렇지 않으면 false
 	 */
 	public Group getGroupByPrimaryKey(String key) {
-		
-		Set set = groups.keySet();
+
+		Set<String> set = groups.keySet();
 		Iterator<String> it = set.iterator();
-		while(it.hasNext()) {
+		while (it.hasNext()) {
 			System.out.print(it.next() + ", ");
 		}
 		System.out.println();
-		
+
 		Group targetGroup = groups.get(key);
 		if (targetGroup != null) {
 		}
@@ -90,7 +94,8 @@ public class Server {
 				for (int i = 0; i < tempFile.length; i++) {
 					if (tempFile[i].isFile()) {
 						tempFile[i].delete();
-					} else { // Recursive function
+					}
+					else { // Recursive function
 						deleteAllFilesInGroupDir(tempFile[i].getPath());
 					}
 					tempFile[i].delete();
@@ -99,10 +104,12 @@ public class Server {
 			}
 		}
 	}
-	
+
 	public void removeGroup(Group group) {
 		Group removeGroup = groups.remove(group.getPrimaryKey());
-		deleteAllFilesInGroupDir(removeGroup.getPrimaryKey());
+		if (removeGroup != null) {
+			deleteAllFilesInGroupDir(removeGroup.getPrimaryKey());
+		}
 	}
 
 	/**
@@ -115,22 +122,22 @@ public class Server {
 	private String generatePrimaryKey(int length) {
 		StringBuffer temp = new StringBuffer();
 		Random rnd = new Random();
-//		for (int i = 0; i < length; i++) {
-//			int rIndex = rnd.nextInt(2) + 1; // 1 또는 2인 랜덤 정수
-//			switch (rIndex) {
-//			// case 0:
-//			// temp.append((char) ((int) (rnd.nextInt(26)) + 65));
-//			// bre;
-//			case 1:
-//				temp.append((char) ((int) (rnd.nextInt(26)) + 97)); //소문자 알파벳
-//				break;
-//			case 2:
-//				temp.append((rnd.nextInt(10))); //0~9 사이의 랜덤 정수
-//				break;
-//			}
-//		}
-//
-//		int rindex = rnd.nextInt(length) + 1; // 1~length 사이의 랜덤 정수
+		// for (int i = 0; i < length; i++) {
+		// int rIndex = rnd.nextInt(2) + 1; // 1 또는 2인 랜덤 정수
+		// switch (rIndex) {
+		// // case 0:
+		// // temp.append((char) ((int) (rnd.nextInt(26)) + 65));
+		// // bre;
+		// case 1:
+		// temp.append((char) ((int) (rnd.nextInt(26)) + 97)); //소문자 알파벳
+		// break;
+		// case 2:
+		// temp.append((rnd.nextInt(10))); //0~9 사이의 랜덤 정수
+		// break;
+		// }
+		// }
+		//
+		// int rindex = rnd.nextInt(length) + 1; // 1~length 사이의 랜덤 정수
 
 		for (int i = 0; i < 2; i++) {
 			temp.append((char) ((int) (rnd.nextInt(26)) + 97)); // 소문자 알파벳
@@ -151,7 +158,8 @@ public class Server {
 				File file = fileList[i];
 				if (file.isFile()) {
 					System.out.println("File name = " + file.getPath());
-				} else if (file.isDirectory()) {
+				}
+				else if (file.isDirectory()) {
 					System.out.println("Dir name = " + file.getPath());
 					subDirList(file.getCanonicalPath().toString());
 				}
@@ -164,5 +172,48 @@ public class Server {
 		Date date = new Date();
 		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd, a hh:mm:ss");
 		System.out.println(sdf.format(date).toString());
+	}
+
+	class ServerStatusCheck extends Thread {
+
+		@Override
+		public void run() {
+			Scanner sc = new Scanner(System.in);
+			String input = null;
+			while (true) {
+				input = sc.nextLine();
+				switch (input) {
+				case "g":
+					System.out.println("┌──────────────SERVER STATUS");
+					System.out.println("├ Group Count: " + groups.size());
+					System.out.println("├ Group List...  ");
+					Set<String> set = groups.keySet();
+					Iterator<String> it = set.iterator();
+					while (it.hasNext()) {
+						System.out.println("│\t- " + it.next());
+					}
+					break;
+
+				default:
+					Group search = groups.get(input);
+					if (search == null) {
+						System.out.println("[SERVER] Group named \"" + input + "\" is not found.");
+					}
+					else {
+						List<String> list = search.getUserList();
+						System.out.println("┌──────────────GROUP INFO: " + input);
+						System.out.println("├ User Count: " + list.size());
+						System.out.println("├ User List...  ");
+						Iterator<String> listIt = list.iterator();
+						while (listIt.hasNext()) {
+							System.out.println("│\t- " + listIt.next());
+						}
+					}
+					break;
+				}
+				System.out.println("└───────────────────────────");
+				input = null;
+			}
+		}
 	}
 }
